Added window swallowing, a la Plan 9's rio.
This feature tracks windows designated as terminal windows, identified
by their window class (by default, this is st windows). When a new
window is mapped, the swallowing feature uses the X Resources Extension
to determine the owning PID of the new window. It then checks to see if
the new window was created by a terminal process.

If a new window is a descendant of a terminal process and that terminal
process hasn't already swallowed another window, the new window is
"swallowed". That is, it takes the place of the terminal window. This
prevents lots of unused terminal windows from cluttering up the
workspace, but still allows command-line navigation and filename
completion. When the swallowed window is closed or destroyed, the
original terminal window is restored in its place.

Note that this feature requires XLib/XCB, as it uses XCB's "res" module
for X Resource Extension access (the version included in vanilla XLib is
too old to be useful). Additionally, it currently works only on Linux,
as platform-specific code is required to walk the process tree to
determine ancestor/descendant relationships.