Removed XCB dependency, added FreeBSD support.
I originally used XCB because it supported the latest version of
the XRes extension, which I was using to get window PIDs. I was
using it to get window PIDs because some processes were setting the
_NET_WM_PID property after mapping and so I wasn't seeing that property.

Now I just did the obvious thing and check for that property after
mapping. This lets us switch back to the older, more portable way
of doing things.